# fix: Improve context gathering performance