--- a/packages/cli/src/fulltext/listProjectFiles.ts
+++ b/packages/cli/src/fulltext/listProjectFiles.ts
@@ -2,79 +2,9 @@ import assert from 'assert';
 import { readdir } from 'fs/promises';
 import { basename, join } from 'path';
 
-export const COMMON_REPO_BINARY_FILE_EXTENSIONS: string[] = [
-  'png',
-  'jpg',
-  'jpeg',
-  'gif',
-  'bmp',
-  'ico',
-  'tiff',
-  'webp',
-  'svg',
-  'mp3',
-  'wav',
-  'ogg',
-  'flac',
-  'aac',
-  'mp4',
-  'webm',
-  'mkv',
-  'avi',
-  'mov',
-  'wmv',
-  'mpg',
-  'flv',
-  'zip',
-  'tar',
-  'gz',
-  'bz2',
-  'xz',
-  '7z',
-  'rar',
-  'pdf',
-  'doc',
-  'docx',
-  'xls',
-  'xlsx',
-  'ppt',
-  'pptx',
-  'odt',
-  'ods',
-  'odp',
-  'rtf',
-  'woff',
-  'woff2',
-  'eot',
-  'ttf',
-  'otf',
-  'ico',
-  'flv',
-  'avi',
-  'mov',
-  'wmv',
-  'mpg',
-  'jar',
-  'war',
-  'ear',
-  'class',
-  'so',
-  'dll',
-  'exe',
-  'min.js',
-  'min.css',
-];
-
 const IGNORE_DIRECTORIES = ['node_modules', 'vendor', 'tmp', 'build', 'dist', 'target'];
 
 const DEFAULT_PROJECT_FILE_LIMIT = 1000;
-
-export function isBinaryFile(fileName: string): boolean {
-  const extension = fileName.split('.').pop();
-  if (extension && COMMON_REPO_BINARY_FILE_EXTENSIONS.includes(extension)) return true;
-
-  return false;
-}
 
 // Produce a modest-sized listing of files in the project.
 // Ignore a standard list of binary file extensions and directories that tend to be full of
@@ -86,7 +16,6 @@ export default async function listProjectFiles(
   const files = new Array<string>();
 
   const ignoreDirectory = (dir: string) => IGNORE_DIRECTORIES.includes(dir);
-  const ignoreFile = (file: string) => isBinaryFile(file);
 
   // Perform a breadth-first traversal of a directory, collecting all non-binary files and
   // applying the directory ignore list.
@@ -101,9 +30,7 @@ export default async function listProjectFiles(
         const fullPath = join(currentDir, entry.name);
         if (entry.isDirectory()) {
           if (!ignoreDirectory(entry.name)) queue.push(fullPath);
-        } else if (entry.isFile()) {
-          if (!ignoreFile(entry.name)) files.push(fullPath);
-        }
+        } else files.push(fullPath);
       }
     }
   };