--- a/packages/cli/tests/unit/fulltext/FileIndex.spec.ts
+++ b/packages/cli/tests/unit/fulltext/FileIndex.spec.ts
@@ -1,5 +1,10 @@
+import { mkdirSync, readdirSync, writeFileSync } from 'node:fs';
+import { join } from 'node:path';
+
 import sqlite3 from 'better-sqlite3';
-import { FileIndex } from '../../../src/fulltext/FileIndex';
+import tmp from 'tmp';
+
+import { FileIndex, filterFiles } from '../../../src/fulltext/FileIndex';
 
 describe('FileIndex', () => {
   let fileIndex: FileIndex;
@@ -61,5 +66,20 @@ describe('FileIndex', () => {
       const results = fileIndex.search(['dir1'], 10);
       expect(results.map((r) => ({ directory: r.directory, fileName: r.fileName }))).toEqual([]);
     });
+  });
+});
+
+describe(filterFiles, () => {
+  it('filters out files with binary extensions, files over 50 kB and non-files', async () => {
+    const dir = tmp.dirSync({ unsafeCleanup: true }).name;
+    writeFileSync(join(dir, 'file.txt'), 'hello');
+    writeFileSync(join(dir, 'file.zip'), 'hello');
+    writeFileSync(join(dir, 'file.json'), 'hello');
+    writeFileSync(join(dir, 'large.txt'), Buffer.alloc(100_000));
+    mkdirSync(join(dir, 'dir'));
+
+    const fileList = readdirSync(dir);
+    const filtered = await filterFiles(dir, fileList);
+    expect(filtered).toEqual(['file.json', 'file.txt']);
   });
 });