--- a/packages/cli/tests/unit/fulltext/listGitProjectFiles.spec.ts
+++ b/packages/cli/tests/unit/fulltext/listGitProjectFiles.spec.ts
@@ -1,6 +1,5 @@
 import * as executeCommandModule from '../../../src/lib/executeCommand';
 import * as utilsModule from '../../../src/utils';
-import { isBinaryFile } from '../../../src/fulltext/listProjectFiles';
 import listGitProjectFiles from '../../../src/fulltext/listGitProjectFIles';
 
 // Mock dependencies
@@ -8,9 +7,6 @@ jest.mock('../../../src/lib/executeCommand');
 jest.mock('../../../src/utils', () => ({
   ...jest.requireActual('../../../src/utils'),
   isFile: jest.fn(),
-}));
-jest.mock('../../../src/fulltext/listProjectFiles', () => ({
-  isBinaryFile: jest.fn(),
 }));
 
 describe('listGitProjectFiles', () => {
@@ -30,17 +26,13 @@ describe('listGitProjectFiles', () => {
       return '';
     });
     jest.mocked(utilsModule.isFile).mockResolvedValue(true);
-    jest
-      .mocked(isBinaryFile)
-      .mockImplementation((fileName) => !fileName.endsWith('.js') && !fileName.endsWith('.ts'));
   });
 
-  it('includes non-binary git-managed files and filters out binary and unmanaged files', async () => {
+  it('includes non-binary git-managed files and filters out unmanaged files', async () => {
     const files = await listGitProjectFiles(mockDirectory);
 
     expect(files).toContain('file1.js');
     expect(files).toContain('newFile.ts');
     expect(files).toContain('modifiedFile.js');
-    expect(files).not.toContain('file2.png'); // filter out binary file
   });
 });