Before (some debugging added):
```
$ time node built/rpc/explain/try.js
collecting context,  ContextCollector {
  appmapDirectories: [ '/home/divide/projects/applandinc.github.io' ],
  sourceDirectories: [ '/home/divide/projects/applandinc.github.io' ],
  vectorTerms: [
    'edit',     'file',
    '+appmap',  'node',
    'diagnose', 'common',
    'problems'
  ],
  charLimit: 15924,
  query: 'edit file +appmap node diagnose common problems'
}
Wrote file index to /tmp/user/1000/appmap-files-1714750213119aYPMUP/index.sqlite
buildFileIndex: 16.683s
Requested char limit: 15924
Collecting context with 5 events per diagram.
collectEvents: 0.253ms
Building source index with 10 files
Indexing 10 files:
  6950 B	_docs/reference/appmap-node.md
   694 B	_plugins/file_exists.rb
  1274 B	_includes/docs/node_support_matrix.html
  1608 B	blog/_posts/2023-06-09-share-your-appmap-files-and-findings-with-your-team.markdown
 22545 B	blog/_posts/2021-02-04-eager-loading-and-the-n-plus-one-query-problem.markdown
   785 B	videos/_posts/2021-04-02-detect-the-n-plus-1-query-problem-in-rails-apps-in-under-5-minutes.markdown
  5609 B	blog/_posts/2022-01-18-why-we-are-bringing-google-maps-for-code-to-node-js-apps.markdown
   268 B	videos/_posts/2021-10-04-how-to-detect-and-fix-the-n-plus-one-problem-for-java-spring-framework-applications.md
   109 B	appmap.yml
576993 B	assets/appmaps/authn_provider_not_logged_in.appmap.json
indexFiles: 28.565s
Wrote file index to /tmp/user/1000/appmap-source-1714750229833CTnfnM/index.sqlite
buildSourceIndex: 28.597s
source index documents:  124
Remaining characters before context: 15924
Characterlimit reached.
Added 16203 characters out of a requested limit of 15924.
Collected an estimated 16203 characters.

real	0m45.678s
user	0m1.857s
sys	0m2.818s
```

After:
```
$ time node built/rpc/explain/try.js
collecting context,  ContextCollector {
  appmapDirectories: [ '/home/divide/projects/applandinc.github.io' ],
  sourceDirectories: [ '/home/divide/projects/applandinc.github.io' ],
  vectorTerms: [
    'edit',     'file',
    '+appmap',  'node',
    'diagnose', 'common',
    'problems'
  ],
  charLimit: 15924,
  query: 'edit file +appmap node diagnose common problems'
}
Wrote file index to /tmp/user/1000/appmap-files-17147501572813UoJhV/index.sqlite
buildFileIndex: 152.792ms
Requested char limit: 15924
Collecting context with 5 events per diagram.
collectEvents: 0.261ms
Building source index with 10 files
Indexing 10 files:
  6950 B	_docs/reference/appmap-node.md
   694 B	_plugins/file_exists.rb
  1274 B	_includes/docs/node_support_matrix.html
  1608 B	blog/_posts/2023-06-09-share-your-appmap-files-and-findings-with-your-team.markdown
 22545 B	blog/_posts/2021-02-04-eager-loading-and-the-n-plus-one-query-problem.markdown
   785 B	videos/_posts/2021-04-02-detect-the-n-plus-1-query-problem-in-rails-apps-in-under-5-minutes.markdown
  5609 B	blog/_posts/2022-01-18-why-we-are-bringing-google-maps-for-code-to-node-js-apps.markdown
   268 B	videos/_posts/2021-10-04-how-to-detect-and-fix-the-n-plus-one-problem-for-java-spring-framework-applications.md
   109 B	appmap.yml
576993 B	assets/appmaps/authn_provider_not_logged_in.appmap.json
indexFiles: 294.845ms
Wrote file index to /tmp/user/1000/appmap-source-1714750157438Qvdzs8/index.sqlite
buildSourceIndex: 343.727ms
source index documents:  124
Remaining characters before context: 15924
Characterlimit reached.
Added 16203 characters out of a requested limit of 15924.
Collected an estimated 16203 characters.

real	0m0.934s
user	0m0.933s
sys	0m0.302s
```

- Disable data safety SQLite features that don't make sense in the temporary databases.
- Insert rows in bulk using a single prepared statement inside a transaction.