@dividedmind 2024-05-03
> By the way I thought it would ignore a file as large as this?
> 
> ```
> 576993 B	assets/appmaps/authn_provider_not_logged_in.appmap.json
> ```

Apparently not...