@appland-release 2024-05-03
:tada: This PR is included in version @appland/appmap-v3.140.1 :tada:

The release is available on:
- `@appland/appmap-v3.140.1`
- [GitHub release](https://github.com/getappmap/appmap-js/releases/tag/%40appland/appmap-v3.140.1)
- `@appland/appmap-v3.140.1`

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: